refactoring

+++ Trains3D/Trains3D/View/TRTrainView.m
@@ -26,7 +26,7 @@ static ODClassType* _TRTrainView_type;
     self = [super init];
     if(self) {
         _smokeView = [TRSmokeView smokeView];
-        _blackMaterial = [EGStandardMaterial standardMaterialWithDiffuse:[EGColorSource applyColor:EGColorMake(0.0, 0.0, 0.0, 1.0)] specularColor:EGColorMake(0.1, 0.1, 0.1, 1.0) specularSize:1.0];
+        _blackMaterial = [EGStandardMaterial standardMaterialWithDiffuse:[EGColorSource applyColor:EGColorMake(0.0, 0.0, 0.0, 1.0)] specularColor:EGColorMake(((float)(0.1)), ((float)(0.1)), ((float)(0.1)), 1.0) specularSize:1.0];
     }
     
     return self;
@@ -38,7 +38,7 @@ static ODClassType* _TRTrainView_type;
 }
 
 - (EGMaterial*)trainMaterialForColor:(EGColor)color {
-    return [EGStandardMaterial standardMaterialWithDiffuse:[EGColorSource applyColor:color] specularColor:EGColorMake(0.3, 0.3, 0.3, 1.0) specularSize:1.0];
+    return [EGStandardMaterial standardMaterialWithDiffuse:[EGColorSource applyColor:color] specularColor:EGColorMake(((float)(0.3)), ((float)(0.3)), ((float)(0.3)), 1.0) specularSize:1.0];
 }
 
 - (void)drawTrains:(id<CNSeq>)trains {
